+++ Libs/libMacGitverCore/RepoMan/Data/BaseData.hpp
